+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CommonOPF Changelog
 
 ## dev on main branch
+- add `Yij_per_unit` and `Yij` admittance matrix getters
 - add `sj(j, net)`, and `sj_per_unit` power getters
 - add `connected_busses(n, net)`
 - make it so `Network[(bus1, bus2)]` _and_ `Network[(bus2, bus1)]` return the edge struct